Fixed a regression issue where excessive memory was allocated for the GLVolumes before sending to the GPU driver. The following commits were partially reverted:
f2defb334dde5d371410ce7da49473d78394e01f Removed GLVolume non-VBO rendering
7a47f9637982061eab2c2c8c6f599c53e069944d GLVolume and GLIndexedVertexArray refactored to send data to gpu at the first render call

Namely, the GLVolume buffers are "shrink to size"'d before sending their
content to the OpenGL driver, and the vertex buffers are populated
as quickly as possible from the GLVolume, so that the same buffer is not
kept twice in RAM on systems, where the RAM is shared with the graphics
card.

Also the memory allocation reporting was improved for the GLVolumes.