Actual fix of the rare crash recently mentioned in dfc31b3
The problem was in fact elsewhere. After an object is added, it is necessary to
update the gizmos. GLGizmoManager::update_data needs to be called, but regardless
of whether any gizmo is currently opened (which is what is_running returns).

To reproduce the crash that this fixes: add an object, delete it, add another one,
open SLA gizmo -> crash