creality.ini: use 120C nozzle temp for bed leveling for CR-6 SE
it seems the CR-6 SE uses 120C for it's strain gauge based bed leveling,
so we'll make this our general default, so we can limit the number of
start_gcode variants we'll need to maintain.

this also introduces a 10 second dwell before any homing happens, to
allow the nozzle to warm up, so that any filament stuck on the nozzle
becomes somewhat pliable.